requirements.txt
streamlit
scikit-learn
pillow
transformers
matplotlib
networkx
google-generativeai
